refactor RankPointPieChart to use user data directly and simplify props
